Refactoring and obsolete code removal
