+++ crates/hyperqueue/src/bin/hq.rs
@@ -30,8 +30,8 @@ use hyperqueue::client::output::outputs::{Output, Outputs};
 use hyperqueue::client::output::quiet::Quiet;
 use hyperqueue::client::status::Status;
 use hyperqueue::common::cli::{get_id_selector, get_task_selector, IdSelectorArg, TaskSelectorArg};
-use hyperqueue::common::fsutils::absolute_path;
 use hyperqueue::common::setup::setup_logging;
+use hyperqueue::common::utils::fs::absolute_path;
 use hyperqueue::dashboard::ui_loop::start_ui_loop;
 use hyperqueue::server::bootstrap::get_client_connection;
 use hyperqueue::transfer::messages::{FromClientMessage, JobInfoRequest, ToClientMessage};
