--- crates/hyperqueue/src/common/controlflowutils.rs
@@ -1,19 +1,0 @@
-#[macro_export]
-macro_rules! get_or_return {
-    ($e:expr) => {
-        match $e {
-            Some(v) => v,
-            _ => return,
-        }
-    };
-}
-
-#[macro_export]
-macro_rules! get_or_return_ok {
-    ($e:expr) => {
-        match $e {
-            Some(v) => v,
-            _ => return Ok(()),
-        }
-    };
-}
