+++ crates/hyperqueue/src/common/placeholders.rs
@@ -13,12 +13,19 @@ use tako::messages::common::{ProgramDefinition, StdioDef};
 use crate::common::parser::NomResult;
 use crate::{JobId, JobTaskId, Map};
 
-// If a new placeholder is added, also change `get_unknown_placeholders`
 pub const TASK_ID_PLACEHOLDER: &str = "TASK_ID";
 pub const JOB_ID_PLACEHOLDER: &str = "JOB_ID";
 pub const INSTANCE_ID_PLACEHOLDER: &str = "INSTANCE_ID";
 pub const CWD_PLACEHOLDER: &str = "CWD";
 pub const SUBMIT_DIR_PLACEHOLDER: &str = "SUBMIT_DIR";
+
+const KNOWN_PLACEHOLDERS: [&str; 5] = [
+    TASK_ID_PLACEHOLDER,
+    JOB_ID_PLACEHOLDER,
+    INSTANCE_ID_PLACEHOLDER,
+    CWD_PLACEHOLDER,
+    SUBMIT_DIR_PLACEHOLDER,
+];
 
 type PlaceholderMap<'a> = Map<&'static str, Cow<'a, str>>;
 
@@ -85,18 +92,10 @@ pub fn fill_placeholders_log(value: &mut PathBuf, job_id: JobId, submit_dir: &Pa
 
 /// Find placeholders in the input that are not supported by HyperQueue.
 pub fn get_unknown_placeholders(input: &str) -> Vec<&str> {
-    let known_placeholders = [
-        TASK_ID_PLACEHOLDER,
-        JOB_ID_PLACEHOLDER,
-        INSTANCE_ID_PLACEHOLDER,
-        CWD_PLACEHOLDER,
-        SUBMIT_DIR_PLACEHOLDER,
-    ];
-
     let mut unknown = Vec::new();
     for placeholder in parse_resolvable_string(input) {
         if let StringPart::Placeholder(placeholder) = placeholder {
-            if !known_placeholders.contains(&placeholder) {
+            if !KNOWN_PLACEHOLDERS.contains(&placeholder) {
                 unknown.push(placeholder);
             }
         }
