+++ crates/hyperqueue/src/common/serverdir.rs
@@ -9,7 +9,7 @@ use serde::de::Error;
 use serde::{Deserialize, Deserializer, Serialize, Serializer};
 
 use crate::common::error::error;
-use crate::common::fsutils::{absolute_path, create_symlink};
+use crate::common::utils::fs::{absolute_path, create_symlink};
 use crate::transfer::auth::{deserialize_key, serialize_key};
 
 #[derive(Clone)]
