+++ crates/hyperqueue/src/common/utils/controlflow.rs
@@ -1,0 +1,9 @@
+#[macro_export]
+macro_rules! get_or_return {
+    ($e:expr) => {
+        match $e {
+            Some(v) => v,
+            _ => return,
+        }
+    };
+}
