+++ crates/hyperqueue/src/common/utils/str.rs
