+++ crates/hyperqueue/src/server/autoalloc/queue/pbs.rs
@@ -1,4 +1,4 @@
-use crate::common::fsutils::get_current_dir;
+use crate::common::utils::fs::get_current_dir;
 use anyhow::Context;
 use std::future::Future;
 use std::path::PathBuf;
@@ -8,7 +8,7 @@ use tako::common::Map;
 
 use crate::common::manager::info::ManagerType;
 use crate::common::manager::pbs::{format_pbs_duration, parse_pbs_datetime};
-use crate::common::timeutils::local_to_system_time;
+use crate::common::utils::time::local_to_system_time;
 use crate::server::autoalloc::queue::common::{
     build_worker_args, check_command_output, create_allocation_dir, create_command, submit_script,
     ExternalHandler,
