+++ crates/tako/src/transfer/mod.rs
@@ -1,5 +1,2 @@
 pub mod auth;
 pub mod transport;
-
-pub type DataConnection =
-    tokio_util::codec::Framed<tokio::net::TcpStream, tokio_util::codec::LengthDelimitedCodec>;
