+++ crates/tako/src/worker/launcher.rs
@@ -106,53 +106,3 @@ pub fn command_from_definitions(definition: &ProgramDefinition) -> crate::Result
 
     Ok(command)
 }
-
-/*pub fn launch_program_from_task(
-    state_ref: WorkerStateRef,
-    task_ref: TaskRef,
-) -> oneshot::Sender<()> {
-    let (sender, receiver) = oneshot::channel();
-    tokio::task::spawn_local(async move {
-        if task_ref.get().is_removed() {
-            // Task was canceled in between start of the task and this spawn_local
-            return;
-        }
-
-        let program = start_program_from_task(&state_ref, &task_ref);
-        tokio::select! {
-            biased;
-            _ = receiver => {
-                let task_id = task_ref.get().id;
-                log::debug!("Launcher cancelled id={}", task_id);
-            }
-            r = program => {
-                match r {
-                    Ok(()) => {
-                        let mut state = state_ref.get_mut();
-                        let task_id = task_ref.get().id;
-                        log::debug!("Program launcher finished id={}", task_id);
-                        state.finish_task(task_ref, 0);
-                        /*
-                        let data_ref = DataObjectRef::new(
-                            task_id,
-                            0,
-                            DataObjectState::Local(LocalData {
-                                serializer: SerializationType::None,
-                                bytes: Default::default(),
-                                subworkers: Default::default()
-                            }),
-                        );
-                        state.add_data_object(data_ref);*/
-                    }
-                    Err(e) => {
-                        log::debug!("Program launcher failed id={}", task_ref.get().id);
-                        let mut state = state_ref.get_mut();
-                        state.finish_task_failed(task_ref, TaskFailInfo::from_string(e.to_string()));
-                    }
-                }
-            }
-        }
-    });
-    sender
-}
-*/
