+++ crates/tako/src/worker/rpc.rs
@@ -26,7 +26,6 @@ use crate::transfer::auth::{
     do_authentication, forward_queue_to_sealed_sink, open_message, seal_message, serialize,
 };
 use crate::transfer::transport::make_protocol_builder;
-use crate::transfer::DataConnection;
 use crate::worker::hwmonitor::HwSampler;
 use crate::worker::launcher::TaskLauncher;
 use crate::worker::reactor::run_task;
@@ -426,194 +425,4 @@ fn resource_allocation_to_msg(
             })
             .collect(),
     }
-}
-
-pub async fn connection_initiator(
-    listener: TcpListener,
-    state_ref: WorkerStateRef,
-) -> crate::Result<()> {
-    loop {
-        let (socket, address) = listener.accept().await?;
-        socket.set_nodelay(true)?;
-        let stream = make_protocol_builder().new_framed(socket);
-
-        let state = state_ref.clone();
-        tokio::task::spawn_local(async move {
-            log::debug!("New connection: {}", address);
-            connection_rpc_loop(stream, state, address)
-                .await
-                .expect("Connection failed");
-            log::debug!("Connection ended: {}", address);
-        });
-    }
-}
-
-async fn connection_rpc_loop(
-    _stream: DataConnection,
-    _state_ref: WorkerStateRef,
-    _address: SocketAddr,
-) -> crate::Result<()> {
-    todo!()
-    /*enum FetchHelperResult {
-        OneShot(tokio::sync::oneshot::Receiver<(SerializationType, Bytes)>),
-        DirectResult(DataResponse, Option<Bytes>),
-    }
-
-    loop {
-        let data = match stream.next().await {
-            None => return Ok(()),
-            Some(data) => data?,
-        };
-        let request: DataRequest = rmp_serde::from_slice(&data)?;
-
-        match request {
-            DataRequest::FetchRequest(msg) => {
-                log::debug!("Object {} request from {} started", msg.task_id, address);
-                let result = {
-                    /* It is now a little bix complex, as we need to get .await
-                       out of borrow of data object and state
-                    */
-                    let maybe_data_ref = state_ref.get().data_objects.get(&msg.task_id).cloned();
-                    let data_ref = match maybe_data_ref {
-                        None => {
-                            log::debug!("Object is not here");
-                            let response = DataResponse::NotAvailable;
-                            stream
-                                .send(rmp_serde::to_vec_named(&response).unwrap().into())
-                                .await?;
-                            continue;
-                        }
-                        Some(x) => x,
-                    };
-                    let mut data_obj = data_ref.get_mut();
-                    match &mut data_obj.state {
-                        DataObjectState::Remote(_) => {
-                            log::debug!("Object is marked as remote");
-                            FetchHelperResult::DirectResult(DataResponse::NotAvailable, None)
-                        }
-                        DataObjectState::InSubworkers(insw) => {
-                            let (sender, receiver) = tokio::sync::oneshot::channel();
-                            data_obj.state = start_local_download(
-                                &mut state_ref.get_mut(),
-                                msg.task_id,
-                                insw,
-                                Subscriber::OneShot(sender),
-                            );
-                            FetchHelperResult::OneShot(receiver)
-                        }
-                        DataObjectState::LocalDownloading(ref mut ld) => {
-                            let (sender, receiver) = tokio::sync::oneshot::channel();
-                            ld.subscribers.push(Subscriber::OneShot(sender));
-                            FetchHelperResult::OneShot(receiver)
-                        }
-                        DataObjectState::Local(local_data) => {
-                            FetchHelperResult::DirectResult(
-                                DataResponse::Data(FetchResponseData {
-                                    serializer: local_data.serializer.clone(),
-                                }),
-                                Some(local_data.bytes.clone()),
-                            )
-                            //let data = rmp_serde::to_vec_named(&).unwrap();
-
-                            /*    Either::Right(
-                            stream.send(data.into()).await?;
-                            stream.send(local_data.bytes.clone()).await?;
-                            log::debug!("Object {} request from {} finished", data_id, address);
-                            continue*/
-                        }
-                        DataObjectState::Removed => unreachable!(),
-                    }
-                };
-
-                let (out_msg, opt_bytes) = match result {
-                    FetchHelperResult::DirectResult(msg, opt_bytes) => (msg, opt_bytes),
-                    FetchHelperResult::OneShot(receiver) => match receiver.await {
-                        Ok((serializer, bytes)) => (
-                            DataResponse::Data(FetchResponseData { serializer }),
-                            Some(bytes),
-                        ),
-                        Err(_) => (DataResponse::NotAvailable, None),
-                    },
-                };
-                let data = serialize(&out_msg).unwrap();
-                stream.send(data.into()).await?;
-                if let Some(bytes) = opt_bytes {
-                    stream.send(bytes).await?;
-                }
-                log::debug!("Object {} request from {} finished", msg.task_id, address);
-            }
-            DataRequest::UploadData(msg) => {
-                log::debug!("Object {} upload from {} started", msg.task_id, address);
-                let data = match stream.next().await {
-                    None => {
-                        log::error!(
-                            "Object {} started to upload but data did not arrived",
-                            msg.task_id
-                        );
-                        return Ok(());
-                    }
-                    Some(data) => data?,
-                };
-                let mut error = None;
-                {
-                    let mut state = state_ref.get_mut();
-                    match state.data_objects.get(&msg.task_id) {
-                        None => {
-                            let data_ref = DataObjectRef::new(
-                                msg.task_id,
-                                data.len() as u64,
-                                DataObjectState::Local(LocalData {
-                                    serializer: msg.serializer,
-                                    bytes: data.into(),
-                                    subworkers: Default::default(),
-                                }),
-                            );
-                            state.add_data_object(data_ref);
-                        }
-                        Some(data_ref) => {
-                            let data_obj = data_ref.get();
-                            match &data_obj.state {
-                                DataObjectState::Remote(_) => {
-                                    /* set the data and check waiting tasks */
-                                    todo!()
-                                }
-                                DataObjectState::InSubworkers(_)
-                                | DataObjectState::LocalDownloading(_) => {
-                                    log::debug!(
-                                        "Uploaded data {} is already in subworkers",
-                                        &msg.task_id
-                                    );
-                                    todo!()
-                                }
-                                DataObjectState::Local(local) => {
-                                    log::debug!(
-                                        "Uploaded data {} is already in worker",
-                                        &msg.task_id
-                                    );
-                                    if local.serializer != msg.serializer
-                                        || local.bytes.len() != data.len()
-                                    {
-                                        log::error!(
-                                            "Incompatible data {} was data uploaded",
-                                            &msg.task_id
-                                        );
-                                        error = Some("Incompatible data was uploaded".into());
-                                    }
-                                }
-                                DataObjectState::Removed => unreachable!(),
-                            }
-                        }
-                    };
-                };
-
-                log::debug!("Object {} upload from {} finished", &msg.task_id, address);
-                let response = DataResponse::DataUploaded(UploadResponseMsg {
-                    task_id: msg.task_id,
-                    error,
-                });
-                let data = serialize(&response).unwrap();
-                stream.send(data.into()).await?;
-            }
-        }
-    }*/
 }
